Moved intValueOf() from Util to StringUtil
